+++ backend/honor-citizen/src/main/java/com/example/honorcitizen/domain/application/dto/ApplicationCreateRequest.java
@@ -54,8 +54,13 @@ public class ApplicationCreateRequest {
         private boolean sameAsApplicant;
         private String name;
         private String phone;
+
+        @NotBlank
         private String zipCode;
+
+        @NotBlank
         private String address;
+
         private String detailAddress;
         private String deliveryRequest;
     }